+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - add: VPK conflict checking system
 - add: feature to delete redundant VPK files for WorkshopVpkAddons
+- add: Workshop VPK Finder
 - improve: better download logic of WorkshopVpkAddon
 - improve: move the addon priority setting to the AddonNode view and support addon global priority (now AddonGroups have priority too)   
 - remove: Flat VPK List (you can enable the Flatten option of the search options to realize this function approximately)